requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0
twilio==8.10.0
openai==1.3.0
anthropic==0.7.0
phonenumbers==8.13.26